LayoutTests/media/modern-media-controls/watchos-media-controls/watchos-media-controls-constructor-expected.txt

Testing the WatchOSMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "watchos media-controls"
PASS mediaControls.layoutTraits is an instance of WatchOSLayoutTraits
PASS mediaControls.state is WatchOSMediaControls.State.Paused

PASS successfullyParsed is true

TEST COMPLETE
